# src/main/resources/swagger-cliente.yaml
openapi: 3.0.3
info:
  title: Crud API automoveis
  version: 1.0.0

paths:
  /inserirclientes:
    post:
      tags:
        - AutoMoveis
      description: Insere um novo auto
      operationId: inserir
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Sucesso
  /removerclientes/{idPath}:
    delete:
      tags:
        - AutoMoveis
      description: Remove um cliente
      operationId: removerClientes
      parameters:
        - $ref: '#/components/parameters/idPath'
      responses:
        '200':
          description: Sucesso
  /atualizarclientes:
    put:
      tags:
        - AutoMoveis
      description: Atualiza um cliente
      operationId: atualizarClientes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Sucesso
  /atualizarenderecos/{idPath}:
    put:
      tags:
        - AutoMoveis
      description: Atualiza um endereco
      operationId: atualizarEnderecos
      parameters:
        - $ref: '#/components/parameters/idPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Endereco'
      responses:
        '200':
          description: Sucesso
  /buscarclientescpf:
    get:
      tags:
        - AutoMoveis
      description: Busca um cliente por cpf
      operationId: buscarClientesCPF
      parameters:
        - $ref: '#/components/parameters/cpf-cliente'
      responses:
        '200':
          description: Sucesso
  /buscarclientes:
    get:
      tags:
        - AutoMoveis
      description: Busca todos os clientes
      operationId: buscarClientes
      responses:
        '200':
          description: Sucesso

components:
  schemas:
    Cliente:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        email:
          type: string
        telefone:
          type: string
        endereco:
            $ref: '#/components/schemas/Endereco'
        carros:
            type: array
            items:
              $ref: '#/components/schemas/Carro'
    Endereco:
      type: object
      properties:
        logradouro:
          type: string
        numero:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        cep:
          type: string
        complemento:
          type: string
    Carro:
      type: object
      properties:
        id:
          type: string
        modelo:
          type: string
        marca:
          type: string
        ano:
          type: string
        cor:
          type: string
        placa:
          type: string
        chassi:
          type: string
        renavam:
          type: string
        combustivel:
          type: string
        kilometragem:
          type: string
        valor:
          type: string
        observacao:
          type: string
        status:
          type: string
        dataCadastro:
          type: string
          format: date
        dataCompra:
          type: string
          format: date
        dataVenda:
          type: string
          format: date

  parameters:
      idPath:
        name: idPath
        in: path
        description: ID do cliente
        required: true
        schema:
          type: string

      cpf-cliente:
        name: cpf-cliente
        in: header
        description: CPF do cliente
        required: true
        schema:
          type: string